third_party/WebKit/LayoutTests/fast/events/pointerevents/touch-pointer-events-expected.txt

CONSOLE ERROR: line 105: Uncaught TypeError: Cannot read property 'pointerId' of undefined
Verifies that tapping on an element fires appropriate pointer events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


--- test PE+TE event sequences without preventDefault ---
green received touchstart at phase=target
lightgreen received touchstart at phase=bubbling
grey received touchstart at phase=bubbling
green received touchmove at phase=target
lightgreen received touchmove at phase=bubbling
grey received touchmove at phase=bubbling
green received touchend at phase=target
lightgreen received touchend at phase=bubbling
grey received touchend at phase=bubbling

--- test PE+TE event sequences with PE preventDefault ---
green received touchstart at phase=target
lightgreen received touchstart at phase=bubbling
grey received touchstart at phase=bubbling
green received touchmove at phase=target
lightgreen received touchmove at phase=bubbling
grey received touchmove at phase=bubbling
green received touchend at phase=target
lightgreen received touchend at phase=bubbling
grey received touchend at phase=bubbling

--- test PE params for single-touch ---
FAIL receivedPEsAtTarget[0].isPrimary should be true. Threw exception TypeError: Cannot read property 'isPrimary' of undefined
FAIL receivedPEsAtTarget[0].button should be 0. Threw exception TypeError: Cannot read property 'button' of undefined
FAIL receivedPEsAtTarget[0].buttons should be 1. Threw exception TypeError: Cannot read property 'buttons' of undefined
FAIL receivedPEsAtTarget[0].clientX should be 148. Threw exception TypeError: Cannot read property 'clientX' of undefined
FAIL receivedPEsAtTarget[0].clientY should be 156. Threw exception TypeError: Cannot read property 'clientY' of undefined
FAIL receivedPEsAtTarget[1].isPrimary should be true. Threw exception TypeError: Cannot read property 'isPrimary' of undefined
FAIL receivedPEsAtTarget[1].button should be 0. Threw exception TypeError: Cannot read property 'button' of undefined
FAIL receivedPEsAtTarget[1].buttons should be 1. Threw exception TypeError: Cannot read property 'buttons' of undefined
FAIL receivedPEsAtTarget[1].clientX should be 168. Threw exception TypeError: Cannot read property 'clientX' of undefined
FAIL receivedPEsAtTarget[1].clientY should be 176. Threw exception TypeError: Cannot read property 'clientY' of undefined
FAIL receivedPEsAtTarget[2].isPrimary should be true. Threw exception TypeError: Cannot read property 'isPrimary' of undefined
FAIL receivedPEsAtTarget[2].button should be 0. Threw exception TypeError: Cannot read property 'button' of undefined
FAIL receivedPEsAtTarget[2].buttons should be 0. Threw exception TypeError: Cannot read property 'buttons' of undefined
FAIL receivedPEsAtTarget.length should be 3. Was 0.

--- test PE isPrimary values ---
FAIL receivedPEsAtTarget[0].isPrimary should be true. Threw exception TypeError: Cannot read property 'isPrimary' of undefined
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
